Add unit test
Add NUnit tests for all the services
Fix name of Review in database (old was ProductReview)

+++ Voam.Core/Services/CartItemService.cs
@@ -36,8 +36,11 @@ namespace Voam.Core.Services
                 .Where(ci => ci.ShoppingCartId == shoopingCartId)
                 .ToListAsync();
 
-            repository.RemoveRange(cartItemsToRemove);
-            await repository.SaveChangesAsync();
+            if (cartItemsToRemove.Any())
+            {
+                repository.RemoveRange(cartItemsToRemove);
+                await repository.SaveChangesAsync();
+            }
         }
     }
 }
